+++ src/actinia_core/rest/ephemeral_processing_with_export.py
@@ -37,7 +37,7 @@ from actinia_core.common.redis_interface import enqueue_job
 from actinia_core.common.process_object import Process
 from actinia_core.common.process_chain import ProcessChainModel
 from actinia_core.common.exceptions import AsyncProcessTermination
-from actinia_core.common.response_models import \
+from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel
 
 __license__ = "GPLv3"
