+++ src/actinia_core/rest/location_management.py
@@ -36,9 +36,9 @@ from actinia_core.common.app import auth
 from actinia_core.common.api_logger import log_api_call
 from actinia_core.rest.user_auth import very_admin_role
 from actinia_core.rest.user_auth import check_user_permissions
-from actinia_core.common.response_models import ProcessingResponseModel
-from actinia_core.common.response_models import SimpleResponseModel
-from actinia_core.common.response_models import MapsetInfoResponseModel
+from actinia_core.models.response_models import ProcessingResponseModel
+from actinia_core.models.response_models import SimpleResponseModel
+from actinia_core.models.response_models import MapsetInfoResponseModel
 from actinia_core.rest.resource_base import ResourceBase
 from actinia_core.rest.persistent_processing import PersistentProcessing
 from actinia_core.rest.mapset_management import PersistentGetProjectionRegionInfo
