+++ src/actinia_core/rest/map_layer_management.py
@@ -32,8 +32,8 @@ from actinia_core.rest.resource_base import ResourceBase
 from actinia_core.common.redis_interface import enqueue_job
 from actinia_core.common.request_parser import glist_parser, extract_glist_parameters
 from actinia_core.common.exceptions import AsyncProcessError
-from actinia_core.common.response_models import ProcessingResponseModel
-from actinia_core.common.response_models import StringListProcessingResultResponseModel
+from actinia_core.models.response_models import ProcessingResponseModel
+from actinia_core.models.response_models import StringListProcessingResultResponseModel
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
