+++ src/actinia_core/rest/mapset_management.py
@@ -42,10 +42,10 @@ from actinia_core.common.redis_interface import enqueue_job
 from actinia_core.common.exceptions import AsyncProcessError
 from actinia_core.rest.user_auth import check_user_permissions
 from actinia_core.rest.user_auth import very_admin_role
-from actinia_core.common.response_models import ProcessingResponseModel, \
+from actinia_core.models.response_models import ProcessingResponseModel, \
     StringListProcessingResultResponseModel, MapsetInfoResponseModel, \
     RegionModel, ProcessingErrorResponseModel
-# from actinia_core.common.response_models import MapsetInfoModel
+# from actinia_core.models.response_models import MapsetInfoModel
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
