+++ src/actinia_core/rest/process_chain_monitoring.py
@@ -33,7 +33,7 @@ from flask import jsonify, make_response, Response
 from flask_restful_swagger_2 import swagger, Schema
 
 from actinia_core.rest.resource_management import ResourceManager
-from actinia_core.common.response_models import SimpleResponseModel
+from actinia_core.models.response_models import SimpleResponseModel
 
 __license__ = "GPLv3"
 __author__ = "Anika Weinmann"
