+++ src/actinia_core/rest/process_validation.py
@@ -32,7 +32,7 @@ from flask import jsonify, make_response
 from actinia_core.rest.ephemeral_processing import EphemeralProcessing
 from actinia_core.rest.resource_base import ResourceBase
 from actinia_core.common.redis_interface import enqueue_job
-from actinia_core.common.response_models import \
+from actinia_core.models.response_models import \
     ProcessingResponseModel, ProcessingErrorResponseModel
 from actinia_core.common.app import auth
 from actinia_core.common.api_logger import log_api_call
