+++ src/actinia_core/rest/raster_legend.py
@@ -33,7 +33,7 @@ from actinia_core.common.redis_interface import enqueue_job
 import tempfile
 import os
 from flask_restful_swagger_2 import swagger
-from actinia_core.common.response_models import ProcessingErrorResponseModel
+from actinia_core.models.response_models import ProcessingErrorResponseModel
 
 
 __license__ = "GPLv3"
