+++ src/actinia_core/rest/raster_renderer.py
@@ -32,7 +32,7 @@ from flask import jsonify, make_response, Response
 from .ephemeral_processing import EphemeralProcessing
 from actinia_core.common.redis_interface import enqueue_job
 from .renderer_base import RendererBaseResource, EphemeralRendererBase
-from actinia_core.common.response_models import ProcessingErrorResponseModel
+from actinia_core.models.response_models import ProcessingErrorResponseModel
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
