+++ src/actinia_core/rest/strds_management.py
@@ -36,7 +36,7 @@ from .persistent_processing import PersistentProcessing
 from .resource_base import ResourceBase
 from actinia_core.common.redis_interface import enqueue_job
 from actinia_core.common.exceptions import AsyncProcessError
-from actinia_core.common.response_models import ProcessingResponseModel, \
+from actinia_core.models.response_models import ProcessingResponseModel, \
     StringListProcessingResultResponseModel, ProcessingErrorResponseModel
 
 __license__ = "GPLv3"
